Remove views.py and moved its comp into other files, fixed unread bug (#67)
* Formatting

* Remove depricated comments

* Removed views.py and shifted dependencies over to correct files

* Changed messages.html to now show read/unread as sending messages will display as read which is not always true. So we can change it to display a shaded colour (using the class) when there isnot any new messageses
