+++ templates/messages.html
@@ -14,8 +14,6 @@
     {% set message_history = msg.message | from_json %}
     <div class="message">
         <button type="submit" name="{{ msg.id }}" class="messageBTNS {% if message_history[-1]['id'] != current_user.id and msg.unread %}unread{% else %}read{% endif %}">
-            <span><strong>{% if message_history[-1]['id'] != current_user.id and msg.unread %} Unread {% else %} Read {% endif %}</strong></span>
-            </br>
             <span><strong>Subject: </strong>{{ msg.subject }}</span>
             </br>
             <span><strong>Sender: </strong>{% if msg.buyer_id == current_user.id %} You {% else %} {{ msg.user.first_name }} {{msg.user.last_name }} {% endif %}</span>
