+++ templates/preview_listing.html
@@ -72,7 +72,7 @@
                 <a class="text-decoration-none text-black button-link">
                     <button type="submit" class="btn btn-primary my-2" name="Share">Share</button>
                 </a>
-                <a href="{{ url_for('views.edit_listing', listing_id=listing.id) }}" class="text-decoration-none text-black button-link">
+                <a href="{{ url_for('private_view.edit_listing', listing_id=listing.id) }}" class="text-decoration-none text-black button-link">
                     <button type="submit" class="btn btn-primary my-2" name="back">Go Back To Edit Listing</button>
                 </a>
             </div>
